requirements.txt
opencv-python
numpy
python-vlc